client_api/requirements.txt
asgiref==3.2.3
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==6.2.2
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
psycopg2==2.8.5
requests==2.23.0
pandas==1.0.3
celery==4.4.2
redis==3.5.2
django-cors-headers==3.3.0
eventlet==0.25.2
python-socketio==4.6.0
uvicorn==0.11.3
aioredis==1.3.1
